move AdminCredentials to shared module
