+++ cerbos/sdk/grpc/client.py
@@ -2,12 +2,12 @@
 # SPDX-License-Identifier: Apache-2.0
 
 from cerbos.sdk._async._grpc import (
-    AdminCredentials,
     AsyncCerbosAdminClient,
     AsyncCerbosClient,
     AsyncPrincipalContext,
 )
 from cerbos.sdk._sync._grpc import CerbosAdminClient, CerbosClient, PrincipalContext
+from cerbos.sdk.grpc.utils import AdminCredentials
 
 __all__ = [
     "AdminCredentials",
