+++ cerbos/sdk/grpc/utils.py
@@ -1,8 +1,25 @@
 # Copyright 2021-2022 Zenauth Ltd.
 # SPDX-License-Identifier: Apache-2.0
 
+from dataclasses import dataclass
+
+from docker.utils.utils import base64
+
 from cerbos.effect.v1.effect_pb2 import EFFECT_ALLOW
 from cerbos.response.v1.response_pb2 import CheckResourcesResponse
+
+
+@dataclass
+class AdminCredentials:
+    username: str = "cerbos"
+    password: str = "cerbosAdmin"
+
+    def metadata(self):
+        credentials = f"{self.username}:{self.password}"
+        encoded_credentials = base64.b64encode(credentials.encode("utf-8")).decode(
+            "utf-8"
+        )
+        return (("authorization", f"Basic {encoded_credentials}"),)
 
 
 def get_resource(
